API: Integrated with livecoding.tv
This is just the first commit but I have created my neccesary endpoints
to allow a user to "authorize" lctva on their behalf (read only for now)
and retrieve an access token, access token, and refresh token.

These tokens will be used for futher API calls.

Time to celebrate :) or... panic.